Extract Discord emoji helpers into a shared utility
