+++ src/discord/interactions/commands/user/handler.ts
@@ -16,8 +16,6 @@ import {
 } from "@/src/discord/interactions/utils/validateInteractions";
 import {
   createDateString,
-  getEmojiStrings,
-  getGameEmojis,
 } from "@/src/discord/discordUtils";
 import {
   calculateLongestStreak,
@@ -25,6 +23,7 @@ import {
 } from "@/src/discord/interactions/utils/streak";
 import { getFastestMessages } from "@/src/discord/interactions/utils/speed";
 import { createMakePublicButton } from "@/src/discord/interactions/components/makePublicButton";
+import { getEmojiStrings, getGameEmojis } from "@/src/discord/utils/emoji";
 
 /**
  * Handles the Discord interaction (slash command) to get user info
